refactor(core): move model helpers into ai-model
